+++ lib/screens/home_screen.dart
@@ -529,7 +529,7 @@ class _HomeScreenState extends State<HomeScreen> {
 
     // Calculate prices and discount display
     final double? productPrice = product.price?.toDouble();
-    final double? productDiscount = product.discount;
+    final int? productDiscount = product.discount;
 
     String displayOriginalPrice =
         '\$${productPrice?.toStringAsFixed(0) ?? '0'}.00';
